Incomplete rid_index/oitg_fetch draft assuming ordered RIDs
The library/command line tools can be used to locate
results files by RID only (and in the case of oitg_fetch,
copy them to the current directory).

This implementation assumed RIDs would be monotonic
across hour directories. This is, unfortunately, not the
case, as the target directories are only set during the
build worker action (invoked during prepare()), not when
the run is created/RID is assigned.

Index size for lab1_alice in November 2020 was 28 kB.